requirements.txt
Flask
Gunicorn
google-generativeai